src/requirements.txt
Django==4.0.5
djangorestframework==3.13.1
drf-yasg==1.20.0
environs==9.5.0
django-filter==22.1
pandas==1.4.3